feat: 项目迁移PostgreSQL并添加人脸识别功能，新增用户注册与登录方式
1. 完成数据库从MySQL到PostgreSQL的迁移，集成pgvector向量扩展
2. 新增人脸识别模块，支持insightface模型加载与人脸特征比对
3. 添加用户注册接口与人脸登录认证流程
4. 重构用户模型，新增人脸特征向量字段与HNSW索引加速检索
5. 前端新增注册、人脸登录页面与相关交互逻辑
6. 清理冗余Alembic迁移文件，优化数据库迁移流程

+++ server/app/models/menu.py
@@ -1,6 +1,5 @@
-from sqlalchemy import String, BigInteger, Integer
+from sqlalchemy import String, BigInteger, Integer, SMALLINT
 from sqlalchemy.orm import Mapped, mapped_column
-from sqlalchemy.dialects.mysql import TINYINT
 from app.models.base import BaseModel
 
 
@@ -46,7 +45,7 @@ class Menu(BaseModel):
         comment="前端组件路径"
     )
     type: Mapped[int] = mapped_column(
-        TINYINT,
+        SMALLINT,
         nullable=False,
         comment="菜单类型：0=目录 1=页面"
     )
@@ -61,22 +60,22 @@ class Menu(BaseModel):
         comment="菜单图标"
     )
     is_hidden: Mapped[int] = mapped_column(
-        TINYINT,
+        SMALLINT,
         default=0,
         comment="是否隐藏：0=显示 1=隐藏"
     )
     is_frame: Mapped[int] = mapped_column(
-        TINYINT,
+        SMALLINT,
         default=0,
         comment="是否内嵌：0=否 1=是"
     )
     is_external: Mapped[int] = mapped_column(
-        TINYINT,
+        SMALLINT,
         default=0,
         comment="是否外部链接：0=否 1=是"
     )
     status: Mapped[int] = mapped_column(
-        TINYINT,
+        SMALLINT,
         default=0,
         comment="状态：0=启用 1=禁用"
     )
